packages/agents/requirements.txt
uagents>=0.12.0
python-dotenv>=1.0.1
requests>=2.32.3
